refactor: rename package gameplay to clue
